Replace graphQl
