+++ Lab3/client/src/components/Todo/Todo.jsx
@@ -7,7 +7,7 @@ const Todo = props => {
     return (
         <div className={Style.Wrapper}>
             <div className={Style.TodoWrapper}>{children}</div>
-            <button onClick={onClick}>
+            <button type="button" onClick={onClick}>
                 &#10006;
             </button>
         </div>
